ft(implement discover page with layout)
